+++ includes/Metaboxes/Exercice/Manager.php
@@ -21,6 +21,7 @@ use ROI\Metaboxes\Exercice\Types\TypeParcours;
 use ROI\Metaboxes\Exercice\Types\TypeClassEchecs;
 use ROI\Metaboxes\Exercice\Types\TypeQuiSuisJe;
 use ROI\Metaboxes\Exercice\Types\TypeOuvreBoite;
+use ROI\Metaboxes\Exercice\Types\TypeCapOuPasCap;
 
 /**
  * Class Manager
@@ -121,6 +122,7 @@ class Manager {
 		( new TypeClassEchecs() )->render( $post, $config_data );
 		( new TypeQuiSuisJe() )->render( $post, $config_data );
 		( new TypeOuvreBoite() )->render( $post, $config_data );
+		( new TypeCapOuPasCap() )->render( $post, $config_data );
 		?>
 
 		<textarea name="roi_exercice_config" id="roi_config_json" style="display:none;"><?php echo esc_textarea( $config ); ?></textarea>